requirements.txt
Flask==2.0.1
pytesseract==0.3.8
Pillow==10.0.0
gunicorn==20.1.0
Werkzeug==2.0.3
